feat: allow timestamp override
There are some cases where a service may wish to override the timestamp - such as when a downstream client is providing the timestamp. This change enables that by using a special `_timestamp` field to denote the new value.

Note that it _must_ be a float64-compatible value, or the type assertion will fail.